+++ crates/tools/hadris-udf-cli/src/commands/info.rs
@@ -2,7 +2,7 @@ use std::fs::File;
 
 use hadris_udf::UdfFs;
 
-use crate::args::InfoArgs;
+use super::super::args::InfoArgs;
 
 use super::Result;