+++ crates/tools/hadris-udf-cli/src/commands/verify.rs
@@ -2,7 +2,7 @@ use std::fs::File;
 
 use hadris_udf::UdfFs;
 
-use crate::args::VerifyArgs;
+use super::super::args::VerifyArgs;
 
 use super::Result;